# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: TravelPlan API
  version: 1.0.0
  description: >-
    # TravelPlan Public API


    Production-grade AI travel management API.


    ## Versioning


    - **Current:** `/api/v1`

    - **Legacy (compatible):** `/api` (same handlers; receives `X-API-Version:
    1`)


    Prefer `/api/v1` for all new integrations.


    ## Authentication


    Most endpoints require a JWT Bearer token from `POST /api/v1/auth/login` or
    signup.


    ```

    Authorization: Bearer <token>

    ```


    Admin monitoring APIs additionally require an admin role (`User.role=admin`
    or `ADMIN_EMAILS`).


    ## Rate limits


    Redis-backed limits (fail-open if Redis is down):


    | Scope | Limit |

    |-------|-------|

    | Login | 10 / minute |

    | Signup | 10 / minute |

    | Forgot password | 5 / minute |

    | AI endpoints | 20 / hour |

    | Public weather | 60 / minute |


    Responses may include `X-RateLimit-*` headers. HTTP **429** when exceeded.


    ## Errors


    Error bodies use:


    ```json

    { "success": false, "code": "ERROR_CODE", "message": "Human readable",
    "error": "..." }

    ```
  contact:
    name: TravelPlan API
    email: hello@travelplan.com
  license:
    name: Proprietary
    identifier: LicenseRef-TravelPlan-Proprietary
servers:
  - url: http://localhost:5000
    description: Local development (paths include /api/v1)
  - url: https://api.travelplan.example
    description: Production host (example)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT from /auth/login or /auth/signup. Send as `Authorization:
        Bearer <token>`."
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        data: {}
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        code:
          type: string
          example: UNAUTHORIZED
        message:
          type: string
        error:
          type: string
    User:
      type: object
      properties:
        _id:
          type: string
          example: 507f1f77bcf86cd799439011
        name:
          type: string
          example: Ada Lovelace
        email:
          type: string
          format: email
          example: ada@example.com
        role:
          type: string
          enum:
            - user
            - admin
          example: user
        createdAt:
          type: string
          format: date-time
    AuthTokenResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        token:
          type: string
          description: JWT access token
        user:
          $ref: "#/components/schemas/User"
...
    SignupRequest:
      type: object
      required:
        - name
        - email
        - password
        - confirmPassword
      properties:
        name:
          type: string
          minLength: 2
          example: Ada Lovelace
        email:
          type: string
          format: email
          example: ada@example.com
        password:
          type: string
          minLength: 6
          example: Secret123!
        confirmPassword:
          type: string
          minLength: 6
          example: Secret123!
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: ada@example.com
        password:
          type: string
          example: Secret123!
...
paths:
...
  /api/v1/auth/signup:
    post:
      summary: Sign up
      description: Sign up
      tags: &a1
        - Authentication
      parameters: &a2 []
      responses: &a3
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokenResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      operationId: signup
      security: &a4 []
      requestBody: &a5
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
  /api/auth/signup:
    post:
      summary: Sign up
      description: |-
        Sign up

        **Legacy alias** of the `/api/v1` path. Prefer versioned URLs.
      tags: *a1
      parameters: *a2
      responses: *a3
      operationId: signup
      security: *a4
      requestBody: *a5
      x-deprecated-prefer: Use /api/v1signup
  /api/v1/auth/login:
    post:
      summary: Login
      description: "Returns JWT. Rate limit: 10/min."
      tags: &a6
        - Authentication
      parameters: &a7 []
      responses: &a8
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokenResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      operationId: login
      security: &a9 []
      requestBody: &a10
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
  /api/auth/login:
    post:
      summary: Login
      description: |-
        Returns JWT. Rate limit: 10/min.

        **Legacy alias** of the `/api/v1` path. Prefer versioned URLs.
      tags: *a6
      parameters: *a7
      responses: *a8
      operationId: login
      security: *a9
      requestBody: *a10
      x-deprecated-prefer: Use /api/v1login
  /api/v1/auth/forgot-password:
    post:
      summary: Forgot password
      description: Forgot password
      tags: &a11
        - Authentication
      parameters: &a12 []
      responses: &a13
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      operationId: forgotPassword
      security: &a14 []
      requestBody: &a15
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
  /api/auth/forgot-password:
    post:
      summary: Forgot password
      description: |-
        Forgot password

        **Legacy alias** of the `/api/v1` path. Prefer versioned URLs.
      tags: *a11
      parameters: *a12
      responses: *a13
      operationId: forgotPassword
      security: *a14
      requestBody: *a15
      x-deprecated-prefer: Use /api/v1forgotPassword
  /api/v1/auth/reset-password/{resetToken}:
    post:
      summary: Reset password
      description: Reset password
      tags: &a16
        - Authentication
      parameters: &a17
        - name: resetToken
          in: path
          required: true
          description: Password reset token
          schema:
            type: string
            example: abc123
      responses: &a18
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      operationId: resetPassword
      security: &a19 []
      requestBody: &a20
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - password
                - confirmPassword
              properties:
                password:
                  type: string
                  minLength: 6
                confirmPassword:
                  type: string
                  minLength: 6
  /api/auth/reset-password/{resetToken}:
    post:
      summary: Reset password
      description: |-
        Reset password

        **Legacy alias** of the `/api/v1` path. Prefer versioned URLs.
      tags: *a16
      parameters: *a17
      responses: *a18
      operationId: resetPassword
      security: *a19
      requestBody: *a20
      x-deprecated-prefer: Use /api/v1resetPassword
  /api/v1/auth/me:
    get:
      summary: Current user
      description: Current user
      tags: &a21
        - Authentication
      parameters: &a22 []
      responses: &a23
        "200":
          description: Current user
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  user:
                    $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      operationId: getCurrentUser
      security: &a24
        - bearerAuth: []
  /api/auth/me:
    get:
      summary: Current user
      description: |-
        Current user

        **Legacy alias** of the `/api/v1` path. Prefer versioned URLs.
      tags: *a21
      parameters: *a22
      responses: *a23
      operationId: getCurrentUser
      security: *a24
      x-deprecated-prefer: Use /api/v1getCurrentUser
...